% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signaculum.R
\name{signaculum}
\alias{signaculum}
\title{Signaculum}
\usage{
signaculum(path, max_age = 60 * 60 * 24 * 365 * 1000)
}
\arguments{
\item{path}{Path to the favicon.}

\item{max_age}{Amount of time before the cache goes stale.}
}
\description{
Use a cached favicon.
}
